+++ services/miner_backend/legacy/download_checkpoint.sh
